+++ src/README.md
@@ -10,8 +10,9 @@
 
 ### Form
 
-- [ ] 이름을 제대로 입력하지 않으면 alert창 (유효하지 않은 이름 길이입니다. 자동차의 이름은 1자이상, 5자 이하만 가능합니다.)이 나온다.
-- [ ] 중복 가능 & 숫자 가능 & 공백가능(공백은 모두 비워준다.) & 특수문자 가능
+- [x] 이름을 제대로 입력하지 않으면 alert창 (유효하지 않은 이름 길이입니다. 자동차의 이름은 1자이상, 5자 이하만 가능합니다.)이 나온다.
+- [x] 중복 가능 & 숫자 가능 & 공백가능(공백은 모두 비워준다.) & 특수문자 가능
+- [ ] 자동차 이름 유효성 검증을 통과하면 하위 시도 UI가 보인다.
 - [ ] 시도할 횟수를 입력하지 않으면 alert창 (입력한 레이싱 횟수가 너무 적습니다. 레이싱 횟수는 1이상이어야 합니다.)이 나온다.
 - [ ] 마지막 부분은 콤마가 없어야 한다.
 - [ ] 유효성 검증에 모두 통과하고 submit하면 Content Component가 보인다.
