--- src/js/Component/Form/index.js
@@ -1,11 +1,0 @@
-import { selector } from "../../util/consts.js";
-import Form from "./Form.js";
-
-
-const render = () => {
-  const app = selector('#app')
-  app.innerHTML = Form()
-}
-
-export default render;
-
