--- src/js/Controller/index.js
@@ -1,2 +1,0 @@
-
-
