+++ src/js/View/index.js
@@ -1,0 +1,17 @@
+import { selector } from "../util/consts.js";
+
+
+export const Component = {
+  fragment: document.createDocumentFragment(),
+  create(component) {
+    this.fragment.textContent = component
+    return this.fragment.textContent
+  }
+}
+
+export const render = (target, component) => {
+  target.innerHTML += component
+}
+
+
+
