+++ src/js/index.js
@@ -1,6 +1,14 @@
-import render from "./Component/Form/index.js";
-import Event from "./Controller/Event/event.js"
+import { RacingFormEvent } from './Controller/Event/event.js';
+import {selector} from './util/consts.js'
+import Form from './View/Component/Form/Form.js';
+import { render } from './View/index.js';
+import { Component } from './View/index.js';
 
-render()
 
-Event()
+const app = selector('#app');
+render(app, Component.create(Form()))
+
+const carNameForm = selector('.car-name-form');
+carNameForm.addEventListener('submit', RacingFormEvent.carNameHandler)
+
+
